Create EditFoodButton.tsx
